+++ rust/cli/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 0.1.2-dev
+
+### Minor
+
+- Do not print the low-confidence warning if the content type was simply overwritten
+
+### Patch
+
+- Update dependencies
+
 ## 0.1.1
 
 ### Minor