+++ rust/lib/CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
 
+## 0.2.0-dev
+
+### Major
+
+- Change `FileType::Ruled` to take `ContentType` directly and remove `RuledType`
+- Change `InferredType::content_type` to describe the content type when overwritten
+- Add `InferredType::inferred_type` for the (possibly overwritten) inferred content type
+
+### Minor
+
+- Add `OverwriteReason` to document why the inferred content type is overwritten
+
+### Patch
+
+- Add inference tests with the new reference files
+- Update features extraction test to the new reference file
+
 ## 0.1.1
 
 ### Minor